Ignore lock files
Emacs creates lock files — symbolic links which contain a process
identifier — when you start modifying a file but haven’t yet saved
it. These would cause serveit to produce a 500 error, because it tries
to stat the referent of these symlinks, which is by design a
nonexistent file.

More generally this will stop the server coming up with a 500 error
whenever the directory structure which is being watched contains a
symbolic link to a file which doesn’t exist.